Start scraping script
Add relevant file paths into the constants file. Implement some utility functions. Most importantly, start the scraping script to  collect the data. So far, it has reached the point of scraping the first few fields from the table (to be completed tomorrow). Also will soon put the data into a sqlite database and that is the scraping script in a nutshell. Strong start