Implement DOCX/PDF export for both modes, add  basic icon
It is now possible to generate DOCX and PDF for both course/cc and general analysis. User can decide what to include in the generation (decent flexibility) and the generation appears robust. The only downside is slightly unstable pdf conversion but there is a warning attached to the button. Also add a basic icon (bad, but better than the tkinter default in terms of relevance). And so, the GUI is pretty much done!! In time for the weekend. May do some final testing and tidying but core functionality done. Need to write up readme and also implement email script too. Hopefully doable by the end of the week, if not, extremely early next week. Omega productive times!!!