Write guides, requirement files, final touches
Detailed guide on the setup and functionality of the project has been written reasonably well that a moderately technical user should be  able to use the project as they wish. The overall GUI itself is somewhat intuitive anyways - nothing too complex is invovled - there is just a lot. Licence selected too - mit as always. That marks the end of a super fast project - despite the speed of this project it still covers a massive scope with many skills. Minimal googling indicates high experience! 2500+ lines done in just over a week - that is probably my new record in how fast I have written a project. Guess this is the type of project you just write super fast because you race against time to complete something substantial in a tiny time - a week and a bit. Great fun little project cementing lots of skills from web scraping, GUI, data processing, graph plotting, table exportation, HCO, email sending. Maybe one-of-a-kind? A nostalgic game. Anyways, moving on...!